django_app/requirements.txt
Django==5.2.4
djangorestframework==3.15.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
pydantic==2.7.1
langchain==0.2.16
langchain-openai==0.1.23
openai==1.51.2
chromadb==0.5.15
pypdf==4.2.0
python-docx==1.1.2
beautifulsoup4==4.12.3
requests==2.32.3
aiohttp==3.9.5